Moves some methods from instructions.h into stapi.h
